More file restructuring.
