+++ src/spline2/CMakeLists.txt
@@ -14,11 +14,11 @@ add_library(spline2_omptarget OBJECT MultiBsplineOffload.cpp)
 target_link_libraries(spline2_omptarget PUBLIC einspline platform_runtime)
 
 if(USE_OBJECT_TARGET)
-  add_library(spline2 OBJECT MultiBspline.cpp)
+  add_library(spline2 OBJECT MultiBspline.cpp SplineUtils.cpp)
 else()
-  add_library(spline2 MultiBspline.cpp)
+  add_library(spline2 MultiBspline.cpp SplineUtils.cpp)
 endif()
-target_link_libraries(spline2 PUBLIC spline2_omptarget)
+target_link_libraries(spline2 PUBLIC spline2_omptarget message qmcio)
 
 if(BUILD_UNIT_TESTS)
   add_subdirectory(tests)
