+++ src/spline2/SplineUtils.cpp
