+++ src/spline2/einspline_util.hpp
@@ -2,7 +2,7 @@
 // This file is distributed under the University of Illinois/NCSA Open Source License.
 // See LICENSE file in top directory for details.
 //
-// Copyright (c) 2016 Jeongnim Kim and QMCPACK developers.
+// Copyright (c) 2026 QMCPACK developers.
 //
 // File developed by: Jeongnim Kim, jeongnim.kim@gmail.com, University of Illinois at Urbana-Champaign
 //                    Mark A. Berrill, berrillma@ornl.gov, Oak Ridge National Laboratory
@@ -12,7 +12,7 @@
 
 
 /** @file einspline_util.hpp
- * @brief utility functions for i/o and bcast of einspline objects
+ * @brief utility functions for bcast gatherv of einspline objects
  *
  */
 #ifndef QMCPLUSPLUS_EINSPLINE_UTILITIES_H
@@ -21,9 +21,6 @@
 #include "Message/CommOperators.h"
 #include "mpi/mpi_datatype.h"
 #include "Host/OutputManager.h"
-#include "OhmmsData/FileUtility.h"
-#include "hdf/hdf_archive.h"
-#include "einspline/multi_bspline_copy.h"
 #include "bspline_traits.hpp"
 #include <limits>
 
@@ -89,80 +86,6 @@ inline void gatherv(Communicate* comm,
     mpi::free_column_type(columntype);
   }
 }
-
-template<unsigned DIM>
-struct dim_traits
-{};
-
-// for 3D multi
-template<>
-struct dim_traits<4>
-{
-  template<typename data_type>
-  static void setdim(data_type& a, hsize_t* dims)
-  {
-    dims[0] = a.spliner.x_grid.num + 3;
-    dims[1] = a.spliner.y_grid.num + 3;
-    dims[2] = a.spliner.z_grid.num + 3;
-    dims[3] = a.spliner.z_stride;
-  }
-};
-
-// for 1D multi
-template<>
-struct dim_traits<2>
-{
-  template<typename data_type>
-  static void setdim(data_type& a, hsize_t* dims)
-  {
-    dims[0] = a.spliner.x_grid.num + 2;
-    dims[1] = a.spliner.x_stride;
-  }
-};
-
-/** specialization of h5data_proxy for einspline_engine
-   */
-template<typename T, unsigned D>
-struct h5data_proxy<einspline_engine<T, D>>
-    : public h5_space_type<T, D + 1>
-{
-  using value_type = T;
-  using Base       = h5_space_type<T, D + 1>;
-  using Base::dims;
-  using Base::get_address;
-  using data_type = einspline_engine<T, D>;
-
-  inline h5data_proxy(const data_type& a) { dim_traits<D + 1>::setdim(a, dims); }
-
-  inline bool read(data_type& ref, hid_t grp, const std::string& aname, hid_t xfer_plist = H5P_DEFAULT)
-  {
-    return h5d_read(grp, aname, get_address(ref.spliner.coefs), xfer_plist);
-  }
-
-  inline bool write(const data_type& ref, hid_t grp, const std::string& aname, hid_t xfer_plist = H5P_DEFAULT) const
-  { return h5d_write(grp, aname.c_str(), Base::rank, dims, get_address(ref.spliner.coefs), xfer_plist); }
-};
-
-namespace einspline
-{
-template<typename IV>
-bool outOfBound(const IV& a)
-{
-  for (int i = 0; i < a.size(); ++i)
-    if (a[i] < 0.0 || a[i] >= 1.0)
-      return true;
-  return false;
-}
-
-template<typename IV>
-bool validRange(const IV& low, const IV& up)
-{
-  bool ok = low[0] < up[0];
-  for (int i = 1; i < low.size(); ++i)
-    ok &= (low[i] < up[i]);
-  return ok;
-}
-} // namespace einspline
 
 } // namespace qmcplusplus
 #endif
